Install VSCode Server on RPi added

+++ resources/install_devops_tools_v2.sh
@@ -89,11 +89,17 @@ mvn -version
 printf "\n\n"
 
 echo "==> Installing VS Code Server"
-wget -q https://raw.githubusercontent.com/chilcano/how-tos/master/resources/vscode_server_install.sh
-chmod +x vscode_server_install.sh 
-. vscode_server_install.sh --arch=$_ARCH --vscs-ver=$_VSCS_VER
-rm -f vscode_server_install.sh 
-printf ">> VS Code Server installed.\n\n"
+
+if [[ "$_ARCH" = "arm" ]]; then
+  
+  curl -fsSL https://code-server.dev/install.sh | sh
+else
+  wget -q https://raw.githubusercontent.com/chilcano/how-tos/master/resources/vscode_server_install.sh
+  chmod +x vscode_server_install.sh 
+  . vscode_server_install.sh --arch=$_ARCH --vscs-ver=$_VSCS_VER
+  rm -f vscode_server_install.sh 
+  printf ">> VS Code Server installed.\n\n"
+fi
 
 echo "==> Loading VS Code Server' settings and installing extensions"
 #wget -q https://raw.githubusercontent.com/chilcano/how-tos/master/resources/vscode_server_setup.sh
